Add Custom Scalar Cursor for paginating results
Connection pattern from GraphQL is used to paginate the results, and
cursor type is a abstraction used by the underliying pagination
implementation to paginate results using MongoDB. Each cursor is track
by a Mongo id represented by ObjectID in mongoose.
A resolver is added inside connection module to represent that ObjectId.
In summary:
  - Add Cursor custom scalar to graphql
  - Add test to added custom scalar
  - Add the custom scalar resolver to server
  - Add mock to cursor scalar